restructure to separate nn components
nn components now show up in a folder caleld nn
also each of them now has its own folder within nn
made changes to CMakeLists.txt to reflect the change
